requirements.txt
Babel==2.7.0
Django==2.2.3
django-cors-headers==3.0.2
django-extensions==2.1.9
django-mysql==3.2.0
django-phonenumber-field==3.0.1
django-rest-auth==0.9.5
django-shell-plus==1.1.7
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
djangorestframework-jwt-refresh-token==0.5
djangorestframework-simplejwt==4.3.0
mysqlclient==1.4.2.post1
phonenumbers==8.10.15
PyJWT==1.7.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
